chore: extract lib/components.tsx file
