+++ packages/react-router/index.tsx
@@ -1,31 +1,35 @@
-import * as React from "react";
-import type {
-  Hash,
-  InitialEntry,
-  Location,
-  MemoryHistory,
-  Path,
-  Pathname,
-  Search,
-  To,
-} from "history";
-import {
-  Action as NavigationType,
-  createMemoryHistory,
-  parsePath,
-  createPath,
-} from "history";
+import type { Hash, Location, Path, Pathname, Search, To } from "history";
+import { Action as NavigationType, parsePath, createPath } from "history";
 
+import type {
+  MemoryRouterProps,
+  NavigateProps,
+  OutletProps,
+  RouteProps,
+  PathRouteProps,
+  LayoutRouteProps,
+  IndexRouteProps,
+  RouterProps,
+  RoutesProps,
+} from "./lib/components";
+import {
+  createRoutesFromChildren,
+  renderMatches,
+  MemoryRouter,
+  Navigate,
+  Outlet,
+  Route,
+  Router,
+  Routes,
+} from "./lib/components";
 import type { Navigator } from "./lib/context";
 import {
   LocationContext,
   NavigationContext,
   RouteContext,
 } from "./lib/context";
-import type { Params, PathMatch, RouteMatch, RouteObject } from "./lib/router";
 import type { NavigateFunction, NavigateOptions } from "./lib/hooks";
 import {
-  _renderMatches,
   useHref,
   useInRouterContext,
   useLocation,
@@ -38,15 +42,12 @@ import {
   useResolvedPath,
   useRoutes,
 } from "./lib/hooks";
+import type { Params, PathMatch, RouteMatch, RouteObject } from "./lib/router";
 import {
   generatePath,
-  invariant,
   matchPath,
   matchRoutes,
-  normalizePathname,
   resolvePath,
-  stripBasename,
-  warning,
 } from "./lib/router";
 
 // re-export from history
@@ -76,313 +77,28 @@ export {
   useRoutes,
 };
 
-///////////////////////////////////////////////////////////////////////////////
-// COMPONENTS
-///////////////////////////////////////////////////////////////////////////////
-
-export interface MemoryRouterProps {
-  basename?: string;
-  children?: React.ReactNode;
-  initialEntries?: InitialEntry[];
-  initialIndex?: number;
-}
-
-/**
- * A <Router> that stores all entries in memory.
- *
- * @see https://reactrouter.com/docs/en/v6/api#memoryrouter
- */
-export function MemoryRouter({
-  basename,
-  children,
-  initialEntries,
-  initialIndex,
-}: MemoryRouterProps): React.ReactElement {
-  let historyRef = React.useRef<MemoryHistory>();
-  if (historyRef.current == null) {
-    historyRef.current = createMemoryHistory({ initialEntries, initialIndex });
-  }
-
-  let history = historyRef.current;
-  let [state, setState] = React.useState({
-    action: history.action,
-    location: history.location,
-  });
-
-  React.useLayoutEffect(() => history.listen(setState), [history]);
-
-  return (
-    <Router
-      basename={basename}
-      children={children}
-      location={state.location}
-      navigationType={state.action}
-      navigator={history}
-    />
-  );
-}
-
-export interface NavigateProps {
-  to: To;
-  replace?: boolean;
-  state?: any;
-}
-
-/**
- * Changes the current location.
- *
- * Note: This API is mostly useful in React.Component subclasses that are not
- * able to use hooks. In functional components, we recommend you use the
- * `useNavigate` hook instead.
- *
- * @see https://reactrouter.com/docs/en/v6/api#navigate
- */
-export function Navigate({ to, replace, state }: NavigateProps): null {
-  invariant(
-    useInRouterContext(),
-    // TODO: This error is probably because they somehow have 2 versions of
-    // the router loaded. We can help them understand how to avoid that.
-    `<Navigate> may be used only in the context of a <Router> component.`
-  );
-
-  warning(
-    !React.useContext(NavigationContext).static,
-    `<Navigate> must not be used on the initial render in a <StaticRouter>. ` +
-      `This is a no-op, but you should modify your code so the <Navigate> is ` +
-      `only ever rendered in response to some user interaction or state change.`
-  );
-
-  let navigate = useNavigate();
-  React.useEffect(() => {
-    navigate(to, { replace, state });
-  });
-
-  return null;
-}
-
-export interface OutletProps {
-  context?: unknown;
-}
-
-/**
- * Renders the child route's element, if there is one.
- *
- * @see https://reactrouter.com/docs/en/v6/api#outlet
- */
-export function Outlet(props: OutletProps): React.ReactElement | null {
-  return useOutlet(props.context);
-}
-
-export interface RouteProps {
-  caseSensitive?: boolean;
-  children?: React.ReactNode;
-  element?: React.ReactNode | null;
-  index?: boolean;
-  path?: string;
-}
-
-export interface PathRouteProps {
-  caseSensitive?: boolean;
-  children?: React.ReactNode;
-  element?: React.ReactNode | null;
-  index?: false;
-  path: string;
-}
-
-export interface LayoutRouteProps {
-  children?: React.ReactNode;
-  element?: React.ReactNode | null;
-}
-
-export interface IndexRouteProps {
-  element?: React.ReactNode | null;
-  index: true;
-}
-
-/**
- * Declares an element that should be rendered at a certain URL path.
- *
- * @see https://reactrouter.com/docs/en/v6/api#route
- */
-export function Route(
-  _props: PathRouteProps | LayoutRouteProps | IndexRouteProps
-): React.ReactElement | null {
-  invariant(
-    false,
-    `A <Route> is only ever to be used as the child of <Routes> element, ` +
-      `never rendered directly. Please wrap your <Route> in a <Routes>.`
-  );
-}
-
-export interface RouterProps {
-  basename?: string;
-  children?: React.ReactNode;
-  location: Partial<Location> | string;
-  navigationType?: NavigationType;
-  navigator: Navigator;
-  static?: boolean;
-}
-
-/**
- * Provides location context for the rest of the app.
- *
- * Note: You usually won't render a <Router> directly. Instead, you'll render a
- * router that is more specific to your environment such as a <BrowserRouter>
- * in web browsers or a <StaticRouter> for server rendering.
- *
- * @see https://reactrouter.com/docs/en/v6/api#router
- */
-export function Router({
-  basename: basenameProp = "/",
-  children = null,
-  location: locationProp,
-  navigationType = NavigationType.Pop,
-  navigator,
-  static: staticProp = false,
-}: RouterProps): React.ReactElement | null {
-  invariant(
-    !useInRouterContext(),
-    `You cannot render a <Router> inside another <Router>.` +
-      ` You should never have more than one in your app.`
-  );
-
-  let basename = normalizePathname(basenameProp);
-  let navigationContext = React.useMemo(
-    () => ({ basename, navigator, static: staticProp }),
-    [basename, navigator, staticProp]
-  );
-
-  if (typeof locationProp === "string") {
-    locationProp = parsePath(locationProp);
-  }
-
-  let {
-    pathname = "/",
-    search = "",
-    hash = "",
-    state = null,
-    key = "default",
-  } = locationProp;
-
-  let location = React.useMemo(() => {
-    let trailingPathname = stripBasename(pathname, basename);
-
-    if (trailingPathname == null) {
-      return null;
-    }
-
-    return {
-      pathname: trailingPathname,
-      search,
-      hash,
-      state,
-      key,
-    };
-  }, [basename, pathname, search, hash, state, key]);
-
-  warning(
-    location != null,
-    `<Router basename="${basename}"> is not able to match the URL ` +
-      `"${pathname}${search}${hash}" because it does not start with the ` +
-      `basename, so the <Router> won't render anything.`
-  );
-
-  if (location == null) {
-    return null;
-  }
-
-  return (
-    <NavigationContext.Provider value={navigationContext}>
-      <LocationContext.Provider
-        children={children}
-        value={{ location, navigationType }}
-      />
-    </NavigationContext.Provider>
-  );
-}
-
-export interface RoutesProps {
-  children?: React.ReactNode;
-  location?: Partial<Location> | string;
-}
-
-/**
- * A container for a nested tree of <Route> elements that renders the branch
- * that best matches the current location.
- *
- * @see https://reactrouter.com/docs/en/v6/api#routes
- */
-export function Routes({
-  children,
-  location,
-}: RoutesProps): React.ReactElement | null {
-  return useRoutes(createRoutesFromChildren(children), location);
-}
-
-///////////////////////////////////////////////////////////////////////////////
-// UTILS
-///////////////////////////////////////////////////////////////////////////////
-
-/**
- * Creates a route config from a React "children" object, which is usually
- * either a `<Route>` element or an array of them. Used internally by
- * `<Routes>` to create a route config from its children.
- *
- * @see https://reactrouter.com/docs/en/v6/api#createroutesfromchildren
- */
-export function createRoutesFromChildren(
-  children: React.ReactNode
-): RouteObject[] {
-  let routes: RouteObject[] = [];
-
-  React.Children.forEach(children, (element) => {
-    if (!React.isValidElement(element)) {
-      // Ignore non-elements. This allows people to more easily inline
-      // conditionals in their route config.
-      return;
-    }
-
-    if (element.type === React.Fragment) {
-      // Transparently support React.Fragment and its children.
-      routes.push.apply(
-        routes,
-        createRoutesFromChildren(element.props.children)
-      );
-      return;
-    }
-
-    invariant(
-      element.type === Route,
-      `[${
-        typeof element.type === "string" ? element.type : element.type.name
-      }] is not a <Route> component. All component children of <Routes> must be a <Route> or <React.Fragment>`
-    );
-
-    let route: RouteObject = {
-      caseSensitive: element.props.caseSensitive,
-      element: element.props.element,
-      index: element.props.index,
-      path: element.props.path,
-    };
-
-    if (element.props.children) {
-      route.children = createRoutesFromChildren(element.props.children);
-    }
-
-    routes.push(route);
-  });
-
-  return routes;
-}
-
-/**
- * Renders the result of `matchRoutes()` into a React element.
- */
-export function renderMatches(
-  matches: RouteMatch[] | null
-): React.ReactElement | null {
-  return _renderMatches(matches);
-}
+// re-export from lib/components.ts
+export type {
+  MemoryRouterProps,
+  NavigateProps,
+  OutletProps,
+  RouteProps,
+  PathRouteProps,
+  LayoutRouteProps,
+  IndexRouteProps,
+  RouterProps,
+  RoutesProps,
+};
+export {
+  createRoutesFromChildren,
+  renderMatches,
+  MemoryRouter,
+  Navigate,
+  Outlet,
+  Route,
+  Router,
+  Routes,
+};
 
 ///////////////////////////////////////////////////////////////////////////////
 // DANGER! PLEASE READ ME!
