+++ packages/react-router/lib/context.ts
@@ -2,7 +2,7 @@ import * as React from "react";
 import type { History, Location } from "history";
 import { Action as NavigationType } from "history";
 
-import { RouteMatch } from "./router";
+import type { RouteMatch } from "./router";
 
 /**
  * A Navigator is a "location changer"; it's how you get to different locations.
